refactor : 메인페이지 코드수정
